Bug 20207: Anonymous Purchase Suggestions show on OPAC even when OPACViewOthersSuggestions is set to Don't Show
When you enable AnonSuggestions and ALLOW patrons that aren't logged in to make purchase suggestions, once a suggestion has successfully been submitted,   Anonymous Purchase Suggestions will show on OPAC even when system preference OPACViewOthersSuggestions, is set to Don't Show.

The suggestions will show on opac-suggestions.pl. However, if you search for the title you suggested, you will get a message saying "You are not authorized to see pending purchase suggestions."

This patch fixes the error, ie if OPACViewOthersSuggestions is set to false not other suggestions are shown (if you are not logged in and AnonSuggestions) OR ONLY your own suggestions if you are logged in

To test:
1) activate AnonSuggestions
2) create an anon suggestion
3) create a suggestion with a logged in user
4) toggle OPACViewOthersSuggestions Show/Don't show
5) all suggestions are shown cgi-bin/koha/opac-suggestions.pl in either case
6) apply patch
7) OPACViewOthersSuggestions is now respected ie suggestions from other users are only shown if activated
8) Please sign off

Sponsored-by: Library of the Natural History Museum Vienna